Add retryable http client and use in FGA module (#356)
* Create retryable http client

This retryable client lives in a new `retryablehttp` package and implements a `Do` method that includes built-in retry logic. It implements exponential backoff and retries on request errors as well as 50X response codes.

* Use retryable http client in FGA module

* Check request context's Done() channel to handle cancellations

* Update minimum delay duration to 250ms

The minimum delay time should be 250ms to account for the initial delay (500ms) with max negative jitter (250ms).